default config update, size_t in Result.h

+++ Result.h
@@ -9,10 +9,10 @@ extern "C" {
 
 // Struct to store results
 typedef struct Result {
-    int cycles;
-    int misses;
-    int hits;
-    int primitiveGateCount;
+    size_t cycles;
+    size_t misses;
+    size_t hits;
+    size_t primitiveGateCount;
 } Result;
 
 #ifdef __cplusplus
